+++ src/commands/index.ts
@@ -17,3 +17,4 @@ export { autoGenerateGroupsCommand } from './auto-generate-groups';
 export { duplicateModelCommand } from './duplicate-model';
 export { cloneRemoteModelCommand } from './clone-remote-model';
 export { extendRemoteModelCommand } from './extend-remote-model';
+export { validateModelCommand } from './validate-model';